# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API REST de Gestión de Vehículos",
        "description": "API RESTful para gestión de vehículos, marcas y propietarios desarrollada con Laravel",
        "contact": {
            "name": "Equipo de Desarrollo",
            "email": "contacto@vehiculos-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
        "/api/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener lista de todas las personas",
                "description": "Listar todas las personas.",
                "operationId": "8cd590e5507a22800e00e2c1287f7433",
                "responses": {
                    "200": {
                        "description": "Lista de personas obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Person"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Crear una nueva persona",
                "description": "Crear una nueva persona.",
                "operationId": "450217f5f71abf751b46caf64cf77a3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "identification_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "María García"
                                    },
                                    "identification_number": {
                                        "type": "string",
                                        "example": "987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Persona creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener una persona específica",
                "description": "Retorna los detalles de una persona por su ID",
                "operationId": "52213fdeeac75168569bb1312277c9e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Persona obtenida exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar una persona existente",
                "description": "Actualiza los datos de una persona",
                "operationId": "2a11738ccccf5ebe6a76d6c8dbea82db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "identification_number"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Carlos Rodríguez"
                                    },
                                    "identification_number": {
                                        "type": "string",
                                        "example": "111222333"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Persona actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Persona actualizada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Person"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Eliminar una persona",
                "description": "Elimina una persona del sistema (también elimina sus asociaciones con vehículos)",
                "operationId": "696ca5da838639a0f383923413f61c82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Persona eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Persona eliminada exitosamente"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/personas/{id}/vehiculos": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener vehículos de una persona",
                "description": "Retorna la lista de vehículos que pertenecen a una persona específica",
                "operationId": "b96f5844ae1f36384f33974c7e9f9165",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la persona",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vehículos de la persona",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vehicle"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Persona no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error en la operación"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error de validación"
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "El campo es obligatorio."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Person": {
                "title": "Persona",
                "description": "Modelo de persona/propietario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "identification_number": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Vehicle": {
                "title": "Vehículo",
                "description": "Modelo de vehículo",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "model": {
                        "type": "string",
                        "example": "Corolla 2024"
                    },
                    "vehicle_brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number_of_doors": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Blanco"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "vehicle_brand": {
                        "$ref": "#/components/schemas/VehicleBrand"
                    },
                    "owners": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Person"
                        }
                    }
                },
                "type": "object"
            },
            "VehicleBrand": {
                "title": "Marca de Vehículo",
                "description": "Modelo de marca de vehículo",
                "properties": {
                    "id": {
                        "description": "ID de la marca",
                        "type": "integer",
                        "example": 1
                    },
                    "brand_name": {
                        "description": "Nombre de la marca",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "country": {
                        "description": "País de origen",
                        "type": "string",
                        "example": "Japón"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-10 10:30:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}